requirements.txt
torch>=2.5.0
torchvision
lightning
opencv-python
pyyaml
kagglehub

transformers>=4.52.0,<=4.52.4
accelerate>=0.34.0,<=1.7.0
pillow>=11.3.0
pandas>=2.3.3
qwen-vl-utils>=0.0.14
tqdm>=4.67.1
wandb>=0.22.2
python-dotenv>=1.1.1
requests>=2.32.5
scikit-learn>=1.3.0  

numpy>=1.23.2,<2.0.0
protobuf>=6.32.1
peft>=0.14.0,<=0.15.2
bitsandbytes>=0.48.1

einops>=0.8.1
huggingface-hub>=0.35.3
jieba>=0.42.1
rouge-chinese>=1.0.3
nltk>=3.8.1
datasets>=2.14.0